+++ doc/releaseNotes.md
@@ -29,3 +29,6 @@ This file describes the main feature changes for released versions of ReadoutCar
 - Added glitchCounter to roc-status (monitoring + JSON output).
 - Added DMA status (enabled / disabled) to roc-status (all output styles, including monitoring).
 - Added FEC counter per link to roc-status (stdout + JSON output).
+
+## next version
+- Added support for CRORC ID where missing: roc-status, roc-config JSON.